🚚 Move Process logs to Dev; rename Logs view → Samples
- Extract <ProcessLogs /> into src/renderer/views/dev-sections so the
  buffer view is dev-only alongside the other diagnostic tooling
- Rename Logs.tsx → Samples.tsx (single-section view, no more
  awkward two-section flex layout)
- Sidebar entry "Logs" → "Samples"; NavId 'logs' → 'samples'
- Wire <ProcessLogs /> between the ReplayLoggingToggle and the
  OverviewInspector in Dev
